third_party/blink/web_tests/wpt_internal/app-history/navigate-event/navigate-appHistory-back-cross-document.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="/common/blank.html"></iframe>
<script>
async_test(t => {
  window.onload = t.step_func(() => {
    i.contentWindow.appHistory.navigate("?foo");
    i.onload = t.step_func(() => {
      i.contentWindow.appHistory.onnavigate = t.step_func_done(e => {
        assert_true(e.cancelable);
        assert_false(e.canRespond);
        assert_false(e.userInitiated);
        assert_false(e.hashChange);
        assert_equals(new URL(e.destination.url).pathname, "/common/blank.html");
        assert_false(e.destination.sameDocument);
        assert_equals(e.formData, null);
        assert_equals(e.info, "hi");
      });
      assert_true(i.contentWindow.appHistory.canGoBack);
      i.contentWindow.appHistory.back({ navigateInfo: "hi" });
    })
  });
}, "navigate event for appHistory.back() - cross-document");
</script>
